move xml tag utilities to xml_tag.cc
they are already declared in src/core/xmltag.h
